update CORS mw to set the correct headers
I originally mistakenly assumed that the Access-Control-Allow-Origin
header could handle a comma separated list of origins, which turned
out to be false. The browser expects only one, and it should match
it's own origin. This meant that I had to refactor this middleware
to be able to tell if the origin was in a list of acceptable origins.

Additionally, I decided to support wildcarded subdomains because
I anticipate arbitrary development deploys for testing and we can
do that by deploying them to randomly gen'd subdomains. Perhaps I
_should_ be saying yagni, but right now I'm saying yolo. If we
decide not to support this, it's simple enough to rip out. Simply
remove the following snippet and update the function comment.

```go
for allowedOrigin := range origins {
	if !strings.HasPrefix(allowedOrigin, "*.") {
		continue
	}

	if strings.HasSuffix(origin, allowedOrigin[1:]) {
		return true
	}
}
```