add exported generic "get" function in http module
this takes advantage of a build-time capability with
create-react-app where it make available any REACT_APP_*
environment variables to the build process, and it can be
referenced via process.env.REACT_APP_*. The environment variable
will be embedded into the transpiled code.

All of that allows us to pass in a different url for the api that
can be used when making api requests. The default is to use the
same origin, which locally is handled via a proxy.