rearrange for functional core imperative shell
Turns out that pretty much all of the functionality is io. This server
is essentially a way to move data from a database into the browser
without any more real business logic. This means there's no real
"functional core" at the moment.

There is now a `shell/` directory which holds this "imperative shell"
logic. If the shell consists of io, then all io related code should live
here. The http package holds all of the protocol level code, and calls
into the app layer. The app layer is supposed to be "the shell" so where
the io -> pure -> io bit is contained. Because storage is io, it lives
in the shell, so the app layer calls into it.

When there is "business logic", presumably it would live in a `core/`
directory, which would hold all of the different packages associated
with a pure function core - no io. The app layer would then call into
this core layer between the io work.

Additionally, I converted some instances to the philosophy of "parse,
don't validate". Creating precise types that exist in packages solely to
support the package api.